refactor(site): fold AgentsPageView into AgentsPageLayout
AgentsPageLayout now renders its sidebar shell and <Outlet/> inline,
matching how every other layout in the codebase renders its own Outlet
without a separate View component. The outlet context type moves into
the layout file.

The stories are converted to mount AgentsPageLayout directly and drive
state through API-level mocks (getChats, models, personal overrides)
instead of View props, and the chat-watch WebSocket is stubbed with an
inert socket.
