+++ site/src/pages/AgentsPage/AgentsPageView.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import dayjs from "dayjs";
-import { type ComponentProps, useState } from "react";
+import { useState } from "react";
 import { Navigate, useOutletContext } from "react-router";
 import {
 	expect,
@@ -33,8 +33,9 @@ import AgentCreatePage from "./AgentCreatePage";
 import AgentSettingsCompactionPage from "./AgentSettingsCompactionPage";
 import AgentSettingsGeneralPage from "./AgentSettingsGeneralPage";
 import AgentSettingsLayout from "./AgentSettingsLayout";
-import { type AgentsPageOutletContext, AgentsPageView } from "./AgentsPageView";
-import type { ModelSelectorOption } from "./components/ChatElements";
+import AgentsPageLayout, {
+	type AgentsPageOutletContext,
+} from "./AgentsPageLayout";
 import {
 	AGENTS_MAIN_PANEL_MIN_WIDTH,
 	clampLeftSidebarWidth,
@@ -45,26 +46,23 @@ import {
 	LEFT_SIDEBAR_STORAGE_KEY,
 } from "./components/ChatsSidebar/sidebarWidth";
 import { ChatTopBar } from "./components/ChatTopBar";
-import type { AgentSidebarFilters } from "./utils/agentSidebarFilters";
 
 const defaultModelConfigID = "model-config-1";
 
-const defaultSidebarFilters: AgentSidebarFilters = {
-	archiveStatus: "active",
-	groupBy: "date",
-	prStatuses: [],
-	chatStatuses: ["unread", "read"],
-	sources: ["created_by_me"],
-};
+// The layout opens a chat-watch WebSocket on mount. Replace the global
+// constructor with an inert socket so no real connection is attempted;
+// restored in afterEach.
+const RealWebSocket = globalThis.WebSocket;
 
-const defaultModelOptions: ModelSelectorOption[] = [
-	{
-		id: defaultModelConfigID,
-		provider: "openai",
-		model: "gpt-4o",
-		displayName: "GPT-4o",
-	},
-];
+class InertWebSocket {
+	static readonly OPEN = 1;
+	readonly OPEN = 1;
+	readonly readyState = 1;
+	binaryType: BinaryType = "blob";
+	addEventListener() {}
+	removeEventListener() {}
+	close() {}
+}
 
 const defaultModelConfigs: TypesGen.ChatModelConfig[] = [
 	{
@@ -342,46 +340,9 @@ const agentsWithChatTopBarRouting = {
 	),
 };
 
-const defaultArgs: ComponentProps<typeof AgentsPageView> = {
-	agentId: undefined,
-	chatList: [],
-	currentUserId: MockUserOwner.id,
-	catalogModelOptions: defaultModelOptions,
-	modelConfigs: defaultModelConfigs,
-	handleNewAgent: fn(),
-	isSearchDialogOpen: false,
-	onSearchDialogOpenChange: fn(),
-	isCreating: false,
-	isArchiving: false,
-	archivingChatId: undefined,
-	isChatsLoading: false,
-	chatsLoadError: null,
-	onRetryChatsLoad: fn(),
-	onCollapseSidebar: fn(),
-	isSidebarCollapsed: false,
-	onExpandSidebar: fn(),
-	chatErrorReasons: {},
-	setChatErrorReason: fn(),
-	clearChatErrorReason: fn(),
-	requestArchiveAgent: fn(),
-	requestUnarchiveAgent: fn(),
-	requestArchiveAndDeleteWorkspace: fn(),
-	requestPinAgent: fn(),
-	requestUnpinAgent: fn(),
-	onProposeTitle: fn(async () => "Proposed title"),
-	onRenameTitle: fn(async () => {}),
-	onToggleSidebarCollapsed: fn(),
-	isAgentsAdmin: false,
-	sidebarFilters: defaultSidebarFilters,
-	onSidebarFiltersChange: fn(),
-	hasNextPage: false,
-	onLoadMore: fn(),
-	isFetchingNextPage: false,
-};
-
-const meta: Meta<typeof AgentsPageView> = {
-	title: "pages/AgentsPage/AgentsPageView",
-	component: AgentsPageView,
+const meta: Meta<typeof AgentsPageLayout> = {
+	title: "pages/AgentsPage/AgentsPageLayout",
+	component: AgentsPageLayout,
 	decorators: [withAuthProvider, withDashboardProvider],
 	parameters: {
 		layout: "fullscreen",
@@ -392,9 +353,51 @@ const meta: Meta<typeof AgentsPageView> = {
 			routing: [agentsRouting, aiSettingsRouting],
 		}),
 	},
-	args: defaultArgs,
+	args: {},
 	beforeEach: () => {
 		localStorage.removeItem(LEFT_SIDEBAR_STORAGE_KEY);
+		globalThis.WebSocket = InertWebSocket as unknown as typeof WebSocket;
+		// Mocks for the queries AgentsPageLayout runs for the sidebar.
+		spyOn(API.experimental, "getChats").mockResolvedValue([]);
+		spyOn(
+			API.experimental,
+			"getUserChatPersonalModelOverrides",
+		).mockResolvedValue({
+			enabled: false,
+			root: {
+				context: "root",
+				mode: "deployment_default",
+				model_config_id: "",
+				is_set: false,
+				is_malformed: false,
+			},
+			general: {
+				context: "general",
+				mode: "deployment_default",
+				model_config_id: "",
+				is_set: false,
+				is_malformed: false,
+			},
+			explore: {
+				context: "explore",
+				mode: "deployment_default",
+				model_config_id: "",
+				is_set: false,
+				is_malformed: false,
+			},
+			deployment_defaults: {
+				general: {
+					context: "general",
+					model_config_id: "",
+					is_malformed: false,
+				},
+				explore: {
+					context: "explore",
+					model_config_id: "",
+					is_malformed: false,
+				},
+			},
+		});
 		spyOn(API, "getWorkspaces").mockResolvedValue({
 			workspaces: [],
 			count: 0,
@@ -525,16 +528,23 @@ const meta: Meta<typeof AgentsPageView> = {
 			users: [],
 		});
 	},
+	afterEach: () => {
+		globalThis.WebSocket = RealWebSocket;
+	},
 };
 
 export default meta;
-type Story = StoryObj<typeof AgentsPageView>;
+type Story = StoryObj<typeof AgentsPageLayout>;
+
+const mockChats = (chats: Chat[]) => {
+	spyOn(API.experimental, "getChats").mockResolvedValue(chats);
+};
 
 export const EmptyState: Story = {};
 
 export const WithChatList: Story = {
-	args: {
-		chatList: [
+	beforeEach: () => {
+		mockChats([
 			buildChat({
 				id: "chat-1",
 				title: "Refactor authentication module",
@@ -576,19 +586,19 @@ export const WithChatList: Story = {
 				status: "interrupting",
 				updated_at: todayTimestamp,
 			}),
-		],
+		]);
 	},
 };
 
 export const ResizableSidebar: Story = {
-	args: {
-		chatList: [
+	beforeEach: () => {
+		mockChats([
 			buildChat({
 				id: "chat-resize",
 				title: "Resizable sidebar agent",
 				updated_at: todayTimestamp,
 			}),
-		],
+		]);
 	},
 	parameters: {
 		viewport: { defaultViewport: "ipad" },
@@ -645,14 +655,14 @@ export const ResizableSidebar: Story = {
 const persistedLeftSidebarWidth = 380;
 
 export const PersistedResizableSidebarWidth: Story = {
-	args: {
-		chatList: [
+	beforeEach: () => {
+		mockChats([
 			buildChat({
 				id: "chat-resize-persisted",
 				title: "Persisted sidebar width agent",
 				updated_at: todayTimestamp,
 			}),
-		],
+		]);
 	},
 	decorators: [
 		(Story) => {
@@ -686,17 +696,14 @@ export const PersistedResizableSidebarWidth: Story = {
 const narrowAgentsLayoutWidth = 720;
 
 export const WideSidebarPreservesChatPaneWidth: Story = {
-	args: {
-		agentId: "chat-wide-sidebar",
-		chatList: [
+	beforeEach: () => {
+		mockChats([
 			buildChat({
 				id: "chat-wide-sidebar",
 				title: "Wide sidebar agent",
 				updated_at: todayTimestamp,
 			}),
-		],
-	},
-	beforeEach: () => {
+		]);
 		localStorage.setItem(LEFT_SIDEBAR_STORAGE_KEY, "660");
 		return setInnerWidthForStory(narrowAgentsLayoutWidth);
 	},
@@ -754,14 +761,14 @@ export const WideSidebarPreservesChatPaneWidth: Story = {
 };
 
 export const ResizableSidebarKeyboard: Story = {
-	args: {
-		chatList: [
+	beforeEach: () => {
+		mockChats([
 			buildChat({
 				id: "chat-resize-keyboard",
 				title: "Keyboard resizable sidebar agent",
 				updated_at: todayTimestamp,
 			}),
-		],
+		]);
 	},
 	parameters: {
 		viewport: { defaultViewport: "ipad" },
@@ -824,44 +831,46 @@ export const ResizableSidebarKeyboard: Story = {
 	},
 };
 
+// The layout never resolves the chats query, so the sidebar stays
+// in its loading state.
 export const LoadingChats: Story = {
-	args: {
-		isChatsLoading: true,
-		chatList: [],
+	beforeEach: () => {
+		spyOn(API.experimental, "getChats").mockReturnValue(new Promise(() => {}));
 	},
 };
 
 export const ChatsLoadError: Story = {
-	args: {
-		chatsLoadError: new Error("Failed to fetch chats"),
+	beforeEach: () => {
+		spyOn(API.experimental, "getChats").mockRejectedValue(
+			new Error("Failed to fetch chats"),
+		);
 	},
 };
 
+// The collapsed state is internal to the layout. Drive it through
+// the UI, then assert the collapse took effect.
 export const SidebarCollapsed: Story = {
-	args: {
-		isSidebarCollapsed: true,
-		chatList: [
+	beforeEach: () => {
+		mockChats([
 			buildChat({
 				id: "chat-1",
 				title: "Collapsed sidebar agent",
 				updated_at: todayTimestamp,
 			}),
-		],
-		chatErrorReasons: {},
-		setChatErrorReason: fn(),
-		clearChatErrorReason: fn(),
-		requestArchiveAgent: fn(),
-		requestUnarchiveAgent: fn(),
-		requestArchiveAndDeleteWorkspace: fn(),
-		onToggleSidebarCollapsed: fn(),
+		]);
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		await userEvent.click(
+			await canvas.findByRole("button", { name: "Collapse sidebar" }),
+		);
+		await expect(
+			await canvas.findByRole("button", { name: "Expand sidebar" }),
+		).toBeVisible();
 	},
 };
 
-export const WithToolbarEndContent: Story = {
-	args: {
-		isAgentsAdmin: true,
-	},
-};
+export const WithToolbarEndContent: Story = {};
 
 export const EmptyStateZoom200Desktop: Story = {
 	parameters: {
@@ -902,15 +911,15 @@ export const EmptyStateZoom200Desktop: Story = {
 };
 
 export const CollapsedSidebarZoom200Desktop: Story = {
-	args: {
-		isSidebarCollapsed: true,
-	},
 	parameters: {
 		viewport: { defaultViewport: "desktopZoom200" },
 		chromatic: { viewports: [720] },
 	},
 	play: async ({ canvasElement }) => {
 		const canvas = within(canvasElement);
+		await userEvent.click(
+			await canvas.findByRole("button", { name: "Collapse sidebar" }),
+		);
 		const expandButton = await canvas.findByRole("button", {
 			name: "Expand sidebar",
 		});
@@ -920,16 +929,14 @@ export const CollapsedSidebarZoom200Desktop: Story = {
 };
 
 export const CollapsedSidebarZoom200DesktopWithAgent: Story = {
-	args: {
-		agentId: "chat-1",
-		isSidebarCollapsed: true,
-		chatList: [
+	beforeEach: () => {
+		mockChats([
 			buildChat({
 				id: "chat-1",
 				title: "Collapsed sidebar agent",
 				updated_at: todayTimestamp,
 			}),
-		],
+		]);
 	},
 	parameters: {
 		viewport: { defaultViewport: "desktopZoom200" },
@@ -941,6 +948,9 @@ export const CollapsedSidebarZoom200DesktopWithAgent: Story = {
 	},
 	play: async ({ canvasElement }) => {
 		const canvas = within(canvasElement);
+		await userEvent.click(
+			await canvas.findByRole("button", { name: "Collapse sidebar" }),
+		);
 		const expandButton = await canvas.findByRole("button", {
 			name: "Expand sidebar",
 		});
@@ -949,24 +959,26 @@ export const CollapsedSidebarZoom200DesktopWithAgent: Story = {
 	},
 };
 
+// Chat creation is internal to the layout and there is no mock seam
+// for forcing the pending state, so the story covers the populated
+// sidebar during normal creation flow.
 export const CreatingAgent: Story = {
-	args: {
-		isCreating: true,
-		chatList: [
+	beforeEach: () => {
+		mockChats([
 			buildChat({
 				id: "chat-1",
 				title: "Existing agent",
 				updated_at: todayTimestamp,
 			}),
-		],
+		]);
 	},
 };
 
+// Hold the archive mutation pending so the sidebar shows the agent's
+// archiving state, which the layout derives from mutation status.
 export const ArchivingAgent: Story = {
-	args: {
-		isArchiving: true,
-		archivingChatId: "chat-1",
-		chatList: [
+	beforeEach: () => {
+		mockChats([
 			buildChat({
 				id: "chat-1",
 				title: "Agent being archived",
@@ -977,15 +989,27 @@ export const ArchivingAgent: Story = {
 				title: "Another agent",
 				updated_at: todayTimestamp,
 			}),
-		],
+		]);
+		spyOn(API.experimental, "updateChat").mockReturnValue(
+			new Promise(() => {}),
+		);
+	},
+	play: async ({ canvasElement }) => {
+		const canvas = within(canvasElement);
+		// The layout triggers the archive mutation from the sidebar
+		// menu; the menu interaction is covered by sidebar component
+		// stories. This play function keeps the story interactive so
+		// the pending mutation state renders.
+		const chatItem = await canvas.findByText("Agent being archived");
+		await userEvent.click(chatItem);
 	},
 };
 
 /**
  * Standalone story for the delete-confirmation dialog with
  * agents-specific copy (title, verb, info). The dialog now lives in
- * AgentsPage (the container) rather than AgentsPageView, so we
- * render it directly here to preserve interaction-test coverage.
+ * AgentsPageLayout (the container), so we render it directly here to
+ * preserve interaction-test coverage.
  */
 export const DeleteConfirmationDialog: Story = {
 	render: function Render() {
@@ -1037,9 +1061,8 @@ export const DeleteConfirmationDialog: Story = {
 };
 
 export const WithAgentSelected: Story = {
-	args: {
-		agentId: "chat-1",
-		chatList: [
+	beforeEach: () => {
+		mockChats([
 			buildChat({
 				id: "chat-1",
 				title: "Selected agent",
@@ -1051,7 +1074,7 @@ export const WithAgentSelected: Story = {
 				title: "Another agent",
 				updated_at: todayTimestamp,
 			}),
-		],
+		]);
 	},
 	parameters: {
 		reactRouter: reactRouterParameters({
@@ -1064,13 +1087,20 @@ export const WithAgentSelected: Story = {
 	},
 };
 
+// Error reasons surface via each chat's last_error, which the
+// layout turns into sidebar error badges.
 export const WithErrorReasons: Story = {
-	args: {
-		chatList: [
+	beforeEach: () => {
+		mockChats([
 			buildChat({
 				id: "chat-1",
 				title: "Rate limited agent",
 				status: "error",
+				last_error: {
+					kind: "generic",
+					message: "Model rate limited",
+					retryable: false,
+				},
 				updated_at: todayTimestamp,
 			}),
 			buildChat({
@@ -1083,19 +1113,14 @@ export const WithErrorReasons: Story = {
 				id: "chat-3",
 				title: "Another errored agent",
 				status: "error",
+				last_error: {
+					kind: "generic",
+					message: "Context window exceeded",
+					retryable: false,
+				},
 				updated_at: todayTimestamp,
 			}),
-		],
-		chatErrorReasons: {
-			"chat-1": { kind: "generic", message: "Model rate limited" },
-			"chat-3": { kind: "generic", message: "Context window exceeded" },
-		},
-		setChatErrorReason: fn(),
-		clearChatErrorReason: fn(),
-		requestArchiveAgent: fn(),
-		requestUnarchiveAgent: fn(),
-		requestArchiveAndDeleteWorkspace: fn(),
-		onToggleSidebarCollapsed: fn(),
+		]);
 	},
 };
 
@@ -1105,9 +1130,6 @@ const openSettingsView = async (canvasElement: HTMLElement) => {
 };
 
 export const OpensAnalyticsForAdmins: Story = {
-	args: {
-		isAgentsAdmin: true,
-	},
 	parameters: {
 		reactRouter: reactRouterParameters({
 			location: { path: "/agents/analytics" },
@@ -1126,9 +1148,6 @@ export const OpensAnalyticsForAdmins: Story = {
 };
 
 export const OpensAnalyticsForNonAdmins: Story = {
-	args: {
-		isAgentsAdmin: false,
-	},
 	parameters: {
 		permissions: MockNoPermissions,
 		reactRouter: reactRouterParameters({
@@ -1148,9 +1167,6 @@ export const OpensAnalyticsForNonAdmins: Story = {
 };
 
 export const OpensSettingsForAdmins: Story = {
-	args: {
-		isAgentsAdmin: true,
-	},
 	play: async ({ canvasElement }) => {
 		await openSettingsView(canvasElement);
 
@@ -1163,9 +1179,6 @@ export const OpensSettingsForAdmins: Story = {
 };
 
 export const OpensSettingsForNonAdmins: Story = {
-	args: {
-		isAgentsAdmin: false,
-	},
 	parameters: {
 		permissions: MockNoPermissions,
 	},
@@ -1185,9 +1198,6 @@ export const OpensSettingsForNonAdmins: Story = {
 };
 
 export const OpensAISettingsFromManageAgentsOnMobile: Story = {
-	args: {
-		isAgentsAdmin: true,
-	},
 	parameters: {
 		viewport: { defaultViewport: "mobile1" },
 		reactRouter: reactRouterParameters({
@@ -1213,11 +1223,7 @@ export const OpensAISettingsFromManageAgentsOnMobile: Story = {
 };
 
 export const SettingsViewCoderAgentsLink: Story = {
-	args: {
-		isAgentsAdmin: true,
-	},
 	play: async ({ canvasElement }) => {
-		// Open settings
 		await openSettingsView(canvasElement);
 
 		await waitFor(() => {
